Move ecmult precomputation code to separate file
